# pos-web/Swagger/pos-api-request.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "POS.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/PaymentMethods": {
      "get": {
        "tags": [
          "PaymentMethods"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PaymentMethods"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.CreatePaymentMethodDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.CreatePaymentMethodDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.CreatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/PaymentMethods/{id}": {
      "get": {
        "tags": [
          "PaymentMethods"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PaymentMethods"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.UpdatePaymentMethodDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.UpdatePaymentMethodDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.UpdatePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.DTOs.PaymentMethod.PaymentMethodDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PaymentMethods"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Pricing": {
      "get": {
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Collections.Generic.IEnumerable`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pricing"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.CreatePricingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.CreatePricingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.CreatePricingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
    "/api/Pricing/{id}": {
      "get": {
        "tags": [
          "Pricing"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pricing"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.UpdatePricingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.UpdatePricingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/POS.Application.DTOs.Pricing.UpdatePricingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[POS.Application.DTOs.Pricing.PricingDto, POS.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pricing"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Soft Delete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/POS.Application.Common.Models.ApiResponse`1[[System.Object, System.Private.CoreLib, Version=9.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]]"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "POS.Application.DTOs.PaymentMethod.CreatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "methodName": {
            "type": "string",
            "nullable": true
          },
          "paymentType": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.PaymentMethod.PaymentMethodDto": {
        "type": "object",
        "properties": {
          "paymentMethodId": {
            "type": "integer",
            "format": "int32"
          },
          "methodName": {
            "type": "string",
            "nullable": true
          },
          "paymentType": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "POS.Application.DTOs.PaymentMethod.UpdatePaymentMethodDto": {
        "type": "object",
        "properties": {
          "paymentMethodId": {
            "type": "integer",
            "format": "int32"
          },
          "methodName": {
            "type": "string",
            "nullable": true
          },
          "paymentType": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "POS.Application.DTOs.Pricing.CreatePricingDto": {
        "type": "object",
        "properties": {
          "productVariantId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "POS.Application.DTOs.Pricing.UpdatePricingDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "format": "double"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your valid token in the text input below.\r\n\r\nExample: \"Bearer eyJhbGciOiJIUzI1Ni...\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}